add the mfg refactor
Same structure that was applied in the diag command is also applied to
the mfg commands too. There is also a specific mfg_common.h that is
used to call common functions that all generation implementations use.
