+++ lib/gen5/mfg_gen5.h
@@ -27,4 +27,33 @@
 
 #include "../switchtec_priv.h"
 
+int switchtec_security_config_get_gen5(struct switchtec_dev *dev,
+				       void *state);
+
+int switchtec_security_config_set_gen5(struct switchtec_dev *dev,
+				       void *setting);
+
+int switchtec_mailbox_to_file_gen5(struct switchtec_dev *dev, int fd);
+
+int switchtec_active_image_index_get_gen5(struct switchtec_dev *dev,
+					  void *index);
+
+int switchtec_active_image_index_set_gen5(struct switchtec_dev *dev,
+					  void *index);
+
+int switchtec_fw_exec_gen5(struct switchtec_dev *dev, int bl2);
+
+int switchtec_boot_resume_gen5(struct switchtec_dev *dev);
+
+int switchtec_sn_ver_get_gen5(struct switchtec_dev *dev, void *info);
+
+int switchtec_secure_state_set_gen5(struct switchtec_dev *dev, int state);
+
+int switchtec_kmsk_set_gen5(struct switchtec_dev *dev, void *public_key,
+			    void *signature, void *kmsk);
+
+int switchtec_read_sec_cfg_file_gen5(struct switchtec_dev *dev,
+				      FILE *setting_file,
+				      void *set);
+
 #endif
