+++ lib/gen6/mfg_gen6.h
@@ -27,4 +27,46 @@
 
 #include "../switchtec_priv.h"
 
+int switchtec_security_config_get_gen6(struct switchtec_dev *dev,
+				       void *state);
+
+int switchtec_security_config_set_gen6(struct switchtec_dev *dev,
+				       void *setting);
+
+int switchtec_mailbox_to_file_gen6(struct switchtec_dev *dev, int fd);
+
+int switchtec_active_image_index_get_gen6(struct switchtec_dev *dev,
+					  void *index);
+
+int switchtec_active_image_index_set_gen6(struct switchtec_dev *dev,
+					  void *index);
+
+int switchtec_fw_exec_gen6(struct switchtec_dev *dev, int bl2);
+
+int switchtec_boot_resume_gen6(struct switchtec_dev *dev);
+
+int switchtec_sn_ver_get_gen6(struct switchtec_dev *dev, void *info);
+
+int switchtec_secure_state_set_gen6(struct switchtec_dev *dev, int state);
+
+int switchtec_kmsk_set_gen6(struct switchtec_dev *dev, void *public_key,
+			    void *signature, void *kmsk);
+
+int switchtec_debug_unlock_gen6(struct switchtec_dev *dev, uint32_t serial,
+				uint32_t ver_sec_unlock, void *public_key,
+				void *signature, void *token);
+
+int switchtec_debug_lock_update_gen6(struct switchtec_dev *dev,
+				     uint32_t serial,
+				     uint32_t ver_sec_unlock,
+				     void *public_key, void *signature);
+
+int switchtec_security_settings_get_gen6(struct switchtec_dev *dev,
+					 void *state);
+
+int switchtec_dbg_unlock_get_token_gen6(struct switchtec_dev *dev, void *token,
+					int token_type);
+
+int switchtec_read_token_file_gen6(FILE *tkn_file, void *token);
+
 #endif
