+++ lib/gen_ops.h
@@ -81,27 +81,27 @@ const struct switchtec_gen_ops switchtec_gen6_ops = {
 	.osa = switchtec_osa_gen6,
 
 	/* Manufacturing */
-	.security_config_get = NULL,
-	.security_config_set = NULL,
-	.mailbox_to_file = NULL,
-	.active_image_index_get = NULL,
-	.active_image_index_set = NULL,
-	.fw_exec = NULL,
-	.boot_resume = NULL,
-	.sn_ver_get = NULL,
-	.secure_state_set = NULL,
-	.kmsk_set = NULL,
-	.debug_unlock = NULL,
-	.debug_lock_update = NULL,
-	.security_settings_get = NULL,
-	.debug_token_unlock_get_token = NULL,
-	.security_state_has_kmsk = NULL,
-	.read_sec_cfg_file = NULL,
-	.read_token_file = NULL,
-	.read_uds_file = NULL,
-	.read_pubk_file = NULL,
-	.read_kmsk_file = NULL,
-	.read_signature_file = NULL,
+	.security_config_get = switchtec_security_config_get_gen6,
+	.security_config_set = switchtec_security_config_set_gen5,
+	.mailbox_to_file = switchtec_mailbox_to_file_gen5,
+	.active_image_index_get = switchtec_active_image_index_get_gen6,
+	.active_image_index_set = switchtec_active_image_index_set_gen6,
+	.fw_exec = switchtec_fw_exec_gen6,
+	.boot_resume = switchtec_boot_resume_gen6,
+	.sn_ver_get = switchtec_sn_ver_get_gen6,
+	.secure_state_set = switchtec_secure_state_set_gen6,
+	.kmsk_set = switchtec_kmsk_set_gen6,
+	.debug_unlock = switchtec_debug_unlock_gen6,
+	.debug_lock_update = switchtec_debug_lock_update_gen4,
+	.security_settings_get = switchtec_security_settings_get_gen6,
+	.debug_token_unlock_get_token = switchtec_dbg_unlock_get_token_gen6,
+	.security_state_has_kmsk = switchtec_security_state_has_kmsk_gen4,
+	.read_sec_cfg_file = switchtec_read_sec_cfg_file_gen5,
+	.read_token_file = switchtec_read_token_file_gen6,
+	.read_uds_file = switchtec_read_uds_file_gen4,
+	.read_pubk_file = switchtec_read_pubk_file_gen4,
+	.read_kmsk_file = switchtec_read_kmsk_file_gen4,
+	.read_signature_file = switchtec_read_signature_file_gen4,
 
 	/* Firmware */
 	.fw_part_id_to_type = NULL,
@@ -167,26 +167,26 @@ const struct switchtec_gen_ops switchtec_gen5_ops = {
 	.osa = switchtec_osa_gen5,
 
 	/* Manufacturing */
-	.security_config_get = NULL,
-	.security_config_set = NULL,
-	.mailbox_to_file = NULL,
-	.active_image_index_get = NULL,
-	.active_image_index_set = NULL,
-	.fw_exec = NULL,
-	.boot_resume = NULL,
-	.sn_ver_get = NULL,
-	.secure_state_set = NULL,
-	.kmsk_set = NULL,
-	.debug_unlock = NULL,
-	.debug_lock_update = NULL,
+	.security_config_get = switchtec_security_config_get_gen5,
+	.security_config_set = switchtec_security_config_set_gen5,
+	.mailbox_to_file = switchtec_mailbox_to_file_gen5,
+	.active_image_index_get = switchtec_active_image_index_get_gen5,
+	.active_image_index_set = switchtec_active_image_index_set_gen5,
+	.fw_exec = switchtec_fw_exec_gen5,
+	.boot_resume = switchtec_boot_resume_gen5,
+	.sn_ver_get = switchtec_sn_ver_get_gen5,
+	.secure_state_set = switchtec_secure_state_set_gen5,
+	.kmsk_set = switchtec_kmsk_set_gen5,
+	.debug_unlock = switchtec_debug_unlock_gen4,
+	.debug_lock_update = switchtec_debug_lock_update_gen4,
 	.security_settings_get = NULL, //NOT Supported
 	.debug_token_unlock_get_token = NULL, //NOT SUPPORTED
-	.security_state_has_kmsk = NULL,
-	.read_sec_cfg_file = NULL,
-	.read_uds_file = NULL,
-	.read_pubk_file = NULL,
-	.read_kmsk_file = NULL,
-	.read_signature_file = NULL,
+	.security_state_has_kmsk = switchtec_security_state_has_kmsk_gen4,
+	.read_sec_cfg_file = switchtec_read_sec_cfg_file_gen5,
+	.read_uds_file = switchtec_read_uds_file_gen4,
+	.read_pubk_file = switchtec_read_pubk_file_gen4,
+	.read_kmsk_file = switchtec_read_kmsk_file_gen4,
+	.read_signature_file = switchtec_read_signature_file_gen4,
 
 	/* Firmware */
 	.fw_part_id_to_type = NULL,
@@ -253,27 +253,27 @@ const struct switchtec_gen_ops switchtec_gen4_ops = {
 	.osa = NULL,
 
 	/* Manufacturing */
-	.security_config_get = NULL,
-	.security_config_set = NULL,
-	.mailbox_to_file = NULL,
-	.active_image_index_get = NULL,
-	.active_image_index_set = NULL,
-	.fw_exec = NULL,
-	.boot_resume = NULL,
-	.sn_ver_get = NULL,
-	.secure_state_set = NULL,
-	.kmsk_set = NULL,
-	.debug_unlock = NULL,
-	.debug_lock_update = NULL,
+	.security_config_get = switchtec_security_config_get_gen4,
+	.security_config_set = switchtec_security_config_set_gen4,
+	.mailbox_to_file = switchtec_mailbox_to_file_gen4,
+	.active_image_index_get = switchtec_active_image_index_get_gen4,
+	.active_image_index_set = switchtec_active_image_index_set_gen4,
+	.fw_exec = switchtec_fw_exec_gen4,
+	.boot_resume = switchtec_boot_resume_gen4,
+	.sn_ver_get = switchtec_sn_ver_get_gen4,
+	.secure_state_set = switchtec_secure_state_set_gen4,
+	.kmsk_set = switchtec_kmsk_set_gen4,
+	.debug_unlock = switchtec_debug_unlock_gen4,
+	.debug_lock_update = switchtec_debug_lock_update_gen4,
 	.security_settings_get = NULL, //Not supported
 	.debug_token_unlock_get_token = NULL, //NOT SUPPORTED
-	.security_state_has_kmsk = NULL,
-	.read_uds_file = NULL,
-	.read_sec_cfg_file = NULL,
-	.read_pubk_file = NULL,
-	.read_kmsk_file = NULL,
-	.read_signature_file = NULL,
-	.dbg_unlock_version_update = NULL,
+	.security_state_has_kmsk = switchtec_security_state_has_kmsk_gen4,
+	.read_uds_file = switchtec_read_uds_file_gen4,
+	.read_sec_cfg_file = switchtec_read_sec_cfg_file_gen4,
+	.read_pubk_file = switchtec_read_pubk_file_gen4,
+	.read_kmsk_file = switchtec_read_kmsk_file_gen4,
+	.read_signature_file = switchtec_read_signature_file_gen4,
+	.dbg_unlock_version_update = switchtec_dbg_unlock_version_update_gen4,
 
 	/* Firmware */
 	.fw_part_id_to_type = NULL,
